Speed up update of fetches by stopping pages with old content
For list that can be fetched by "sort=updated&direction=desc", we stop
fetching pages when the last item of the last fetched page is older than
the last fetch of this list. It allows a really big speed up for big
repositories: not need to load 50 pages of issues if only one issue was
modified !

There is one problem: if this limit is reached, we aren't able to check
for deleted entries. But as only issues and comments are affected by
this, and as issues cannot be deleted, only comments, we actually
consider it's ok...