Save PRs even if repo has issues disabled
We don't use the issues.github_id field as an identifier, but
repo+number because issues and PRs don't share the same id on the github
side, and in repositories with issues disabled, we don't have any issue
at all, only PRs.
Note that we have a problem with events, has issues events cannot be
fetched when issues are disabled so we don't have events for the pull
requests... i send a mail to github about this.

Close #12