Reduce the fields to manage from the API sooner
Now the github_ignore attribute of the fill is filled to avoid maxium
calls to "get_field_by_name". And check to ignore a name is done before
checking for its maching: we really test the github keys.
We also remove directly keys starting by '_' or ending by 'fetched_at'
and 'etag'.