Minimize a race-condition when posting/editing
Steps:
1/ the comment is posted by the user
2/ a task is created
3/ the task starts running
3.1/ the comment if fetched via its ID
3.2/ the comment is pushed to github
3.3/ the comment is updated locally
3.4/ a task to fetch the issue is created
4/ the task ends

Before this commit, if a(nother) fetch of the issue is done between 3.1
and 3.4 (maybe via a hook just received), 3.4 fail because "self.object"
refers to a comment that does not exist anymore.
Why ?
Because while fetching the issue, the comment just created, not in the
just fetched data, is deleted. And in 3.3, a *new* comment is created,
with a new ID.
In can happen easily if two people are discussing an issue at the same
time.

By reloa the last "self.object" by "comment", to access its issue, it
works because the comment does not need to be fetched from the DB again
and we don't care of it, we just want the issue.

Note that other race conditions can happen, for example if the parallel
fetch occurs during steps 1 or 2, all created or edited comments for the
fetched issues will be vanished.

A solution could be to only delete *already fetched* objects when
deleting during a fetch.