Add a "clone" method to our Job model
Fields to copy must be specified in the `clonable_fields` attribute of
the Job.
The new job can be delayed by passing `delayed_for` or `delayed_until`
to the  `clone` method. If none, the new Job will be queued to be run
ASAP.
The property of the new job will be `0`, but it can be passed to the
`clone` method too.