Better tracking of label updates
Because they are a M2M they may be saved after the save of the issue, so
the tracker won't see them.
Corrected now to track m2m changes too, and with a hack of the m2m
descriptor of django to remove removed entries, then add new ones,
instead of clearing all then adding.

Fixed #67