Keep commits removed from PRs
When a pull request is rebased, the olds commits are not tied to this
pull request anymore. But if there were comments attached to theses
commits, we want to display them, so we must keep the relation issue-
commit.

Know, this relation is done not via s simple m2m but via a through
table, with a new attribute, 'deleted', that is True if the commit is
not anymore tied to the issue.

cf #28